requirements.txt
numpy
scikit-image
pycrypto
